% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DoCorPoint}
\alias{DoCorPoint}
\title{Visualize the correlation of two genes}
\usage{
DoCorPoint(expr, gene.x, gene.y, method = "pearson")
}
\arguments{
\item{expr}{Expression matrix, rows are genes and columns are cells}

\item{gene.x}{Genes displayed on the x-axis}

\item{gene.y}{Genes displayed on the y-axis}

\item{method}{a character string indicating which correlation coefficient is
to be used for the test. One of "pearson", "kendall", or "spearman",
can be abbreviated.}
}
\value{
a list of ggplot objects
}
\description{
Visualize the correlation of two genes
}
