% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello(string = "Hello, world!")
}
\arguments{
\item{string}{Any character, the default is "Hello, world!".}
}
\value{
Print out the string.
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
You can learn more about package authoring with RStudio at:
    http://r-pkgs.had.co.nz/
Some useful keyboard shortcuts for package authoring:
   Install Package:           'Ctrl + Shift + B'
   Check Package:             'Ctrl + Shift + E'
   Test Package:              'Ctrl + Shift + T'
}
\examples{
hello("Good, job!")
}
