% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna.R
\name{scibet}
\alias{scibet}
\title{SciBet: Single Cell Identifier Based on Entropy Test}
\usage{
scibet(object, model.path, assay = "RNA", slot = "data")
}
\arguments{
\item{object}{An Seurat object}

\item{model.path}{A csv format file, pre-trained prediction model path}

\item{assay}{Assay to use in cell annotation}

\item{slot}{Slot to pull data from}
}
\value{
The vector contains the barcode of each cell and predicted cell type
}
\description{
SciBet: Single Cell Identifier Based on Entropy Test
}
\examples{
\dontrun{
# Import Seurat objects
Load("pbmc_small")
# Cell type prediction
label <- scibet(seurat.obj, "/public/major_human_cell_types.csv")
# Add the prediction result to the Seurat object
seurat.obj$label <- label
}
}
