% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scihub-dl.R
\name{scihub_dl}
\alias{scihub_dl}
\title{Download publication via Sci-Hub}
\usage{
scihub_dl(doi, scihub = "sci-hub.se", download = TRUE, mode = "wb")
}
\arguments{
\item{doi}{The publication doi you want to download}

\item{scihub}{Available URLs of Sci-Hub}

\item{download}{Whether to download the PDF file of the publication}

\item{mode}{The mode with which to write the file. Useful values are "w", "wb" (binary), "a" (append) and "ab". Not used for methods "wget" and "curl". See also ‘Details’, notably about using "wb" for Windows.}
}
\value{
PDF file URL address and PDF file
}
\description{
Use Sci-Hub to download publications via doi
}
\examples{
\dontrun{
scihub_dl("10.1016/j.carbpol.2020.116446")
}
}
